+++ spec/fixtures/scanner_app/app/models/post.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class Post
+  def published?
+    true
+  end
+
+  def self.recent
+    []
+  end
+end